pyproject.toml
[project]
name = "capacitacao"
version = "0.1.0"
description = ""
authors = [
    {name = "Renato",email = "renato.costa87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "kedro[jupyter] (>=1.0.0,<2.0.0)",
    "ipython (==8.0)",
    "jupyterlab (>=4.4.7,<5.0.0)",
    "kedro-datasets (>=8.1.0,<9.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "kedro-viz (>=12.1.0,<13.0.0)",
    "scikeras (>=0.13.0,<0.14.0)",
    "shap (>=0.49.1,<0.50.0)",
    "tensorflow-cpu (>=2.20.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
